Stop flickering by stop syncing
On slower terminals (SSH or locally, zoomed way out), rendering of the UI would flicker regularly. This was a result of excessive calls to `termui.TermHeight()` and `termui.TermWidth()` which internally calls `termbox.Sync()`. From the docs (https://godoc.org/github.com/nsf/termbox-go#Sync): "Sync forces a complete resync between the termbox and a terminal, it may not be visually pretty though."

To prevent this, the width and height of the terminal is only requested at initial startup and in the event of a terminal resize. Great success!