Make Packet Streaming Great Again
tl;dr - if nothing is bound to the sniffed port, you won't see all the packets.

Noticed at low metric rates (10 rps or so), ~50% of the packets weren't picked up by pcap. At higher rates (1k rps), ~10% of packets were not showing up. gopacket reported no packet loss. After much hair-pulling, I started up StatsD to see if perhaps my network was to blame ... and serendipitously all the metrics began to be captured!

This patch aims to eliminate this issue for the user so that they can use statstee with or without a daemon listening. By default now, statstee will first attempt to bind to the port (default: 8125), if it fails, then it will use pcap to capture the metrics. This behavior can overridden using the `-l` or `-c` flags at startup to force either listener or capture modes.