Jerks
Adds the `jerks` example, a load testing utility for statstee. Starting jerks you can specify the number of metrics and the rate at which those metrics should be generated and sent over the wire. This tool will be used to refine the pipeline to efficiently and accurately capture over a large range of metric volumes.